services/booking/api/requirements.txt
pydantic==1.10.2
fastapi==0.92.0
uvicorn==0.20.0
orjson==3.8.7
faker==17.6.0
pyjwt==2.6.0
requests==2.28.2
aiohttp==3.8.1
asyncio==3.4.3
sqlalchemy==1.4.45
asyncpg==0.27.0
psycopg2-binary==2.9.5
databases==0.7.0
aioredis==2.0.1
pytz==2023.3
gunicorn==20.1.0
python-logstash==0.4.8
redis==4.5.4
faker==17.6.0
sentry-sdk==1.23.1
